+++ runtime/vm/compiler/jit/compiler.cc
@@ -174,6 +174,7 @@ void IrregexpCompilationPipeline::ParseFunction(
   RegExpParser::ParseRegExp(pattern, multiline, compile_data);
 
   regexp.set_num_bracket_expressions(compile_data->capture_count);
+  regexp.set_capture_name_map(compile_data->capture_name_map);
   if (compile_data->simple) {
     regexp.set_is_simple();
   } else {