+++ sdk/lib/_internal/js_runtime/lib/js_helper.dart
@@ -45,7 +45,11 @@ import 'dart:_foreign_helper'
 import 'dart:_interceptors';
 import 'dart:_internal' as _symbol_dev;
 import 'dart:_internal'
-    show EfficientLengthIterable, MappedIterable, IterableElementError;
+    show
+        EfficientLengthIterable,
+        MappedIterable,
+        IterableElementError,
+        SubListIterable;
 
 import 'dart:_native_typed_data';